Remove this unnecessary cast to int issue

+++ src/main/java/org/wpsim/marketplace/data/MarketPlaceMessage.java
@@ -1,11 +1,11 @@
 /**
  * ==========================================================================
- * __      __ _ __   ___  * WellProdSim                                  *
- * \ \ /\ / /| '_ \ / __| * @version 1.0                                 *
- * \ V  V / | |_) |\__ \ * @since 2023                                  *
- * \_/\_/  | .__/ |___/ * *
- * | |          * @author Jairo Serrano                        *
- * |_|          * @author Enrique Gonzalez                     *
+ * __      __ _ __   ___  *    WellProdSim                                  *
+ * \ \ /\ / /| '_ \ / __| *    @version 1.0                                 *
+ * \ V  V / | |_) |\__ \ *    @since 2023                                  *
+ * \_/\_/  | .__/ |___/ *                                                 *
+ * | |          *    @author Jairo Serrano                        *
+ * |_|          *    @author Enrique Gonzalez                     *
  * ==========================================================================
  * Social Simulator used to estimate productivity and well-being of peasant *
  * families. It is event oriented, high concurrency, heterogeneous time     *
@@ -58,7 +58,7 @@ public class MarketPlaceMessage extends DataBESA {
 
     public MarketPlaceMessage(MarketPlaceMessageType marketPlaceMessageType, int quantity, String currentDate) {
         this.marketMessageType = marketPlaceMessageType;
-        this.quantity = quantity; // Cast innecesario (int) eliminado
+        this.quantity = (int) quantity;
         this.currentDate = currentDate;
     }
